refactor: extract ArgumentPropsReviewer to own component
